# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "MeasureMetadata": {
                "properties": {
                    "metadata": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object"
                    },
                    "ref": {
                        "type": "string"
                    }
                },
                "required": [
                    "ref",
                    "metadata"
                ]
            },
...
            "OrderedColumn": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "order": {
                        "enum": [
                            "Asc",
                            "Desc"
                        ],
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "order"
                ]
            },
...
            "Predicate": {
                "properties": {}
            },
...
            "QueryReqWithRefs": {
                "properties": {
                    "aggregateFilter": {
                        "$ref": "#/components/schemas/Predicate"
                    },
                    "columns": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "columnsTopN": {
                        "$ref": "#/components/schemas/TopN"
                    },
                    "customReference": {
                        "type": "string"
                    },
                    "filter": {
                        "$ref": "#/components/schemas/Predicate"
                    },
                    "limit": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "measures": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "offset": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "orderBy": {
                        "items": {
                            "$ref": "#/components/schemas/OrderedColumn"
                        },
                        "type": "array"
                    },
                    "queryId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "rows": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "rowsTopN": {
                        "$ref": "#/components/schemas/TopN"
                    },
                    "sql": {
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            "Aggregate",
                            "Record",
                            "Sql",
                            "Native"
                        ],
                        "type": "string"
                    }
                },
                "required": [
                    "type"
                ]
            },
            "Ref": {
                "properties": {
                    "ref": {
                        "type": "string"
                    }
                },
                "required": [
                    "ref"
                ]
            },
            "Response": {
                "properties": {
                    "columns": {
                        "items": {
                            "$ref": "#/components/schemas/ResponseMetaData"
                        },
                        "type": "array"
                    },
                    "measureHeaders": {
                        "items": {
                            "$ref": "#/components/schemas/Ref"
                        },
                        "type": "array"
                    },
                    "measures": {
                        "items": {
                            "$ref": "#/components/schemas/MeasureMetadata"
                        },
                        "type": "array"
                    },
                    "rows": {
                        "items": {
                            "items": {
                                "type": "string"
                            },
                            "type": "array"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "columns",
                    "rows",
                    "measureHeaders",
                    "measures"
                ]
            },
            "ResponseMetaData": {
                "properties": {}
            },
...
            "TopN": {
                "properties": {
                    "dimensions": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "n": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "orderBy": {
                        "items": {
                            "$ref": "#/components/schemas/OrderedColumn"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "orderBy",
                    "dimensions",
                    "n"
                ]
            },
...
        }
    },
    "info": {
        "description": "Pantheon REST API",
        "title": "Pantheon API",
        "version": "0.11.0"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/catalogs/{catalogId}/schemas/{schemaId}/query": {
            "post": {
                "operationId": "executeQuery",
                "parameters": [
                    {
                        "in": "path",
                        "name": "catalogId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "schemaId",
                        "required": true,
                        "schema": {
                            "format": "uuid",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryReqWithRefs"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "success"
                    }
                },
                "summary": "Run query on schema",
                "tags": [
                    "Queries"
                ]
            }
        },
...
    },
...
}